Fix timer sync issues with back button
This change seems to fix the problem. For whatever reason, when you
return to the page via the back button, the clock is not updated to
the current time (seems to be set to the time the user left the
page).  This causes the clock sync to produce an incorrect offset.
This change seems to cause the offset to get set correctly.

Fixes #16